Move SessionPurposeAndTypeField / -Widget into a fields.py file
